moved masks_to_boxes into boxes module
